[MS] Moved a few components around
