+++ client/src/components/core/ms-input/MsChoosePasswordInput.vue
