+++ client/tests/component/specs/testOrganizationCard.spec.ts
@@ -1,6 +1,6 @@
 // Parsec Cloud (https://parsec.cloud) Copyright (c) BUSL-1.1 (eventually AGPL-3.0) 2016-present Scille SAS
 
-import OrganizationCard from '@/components/home/OrganizationCard.vue';
+import OrganizationCard from '@/components/organizations/OrganizationCard.vue';
 import { AvailableDevice } from '@/plugins/libparsec/definitions';
 import { VueWrapper, mount } from '@vue/test-utils';
